bug fixes IOFactory added colorProperty option to PoinSetOpen3D bug fixes RasterData bug fixes mumford shaw elements. Now flow accommodate Chan-Vese bug fixes LevelSetFunction drawContours now flips the image

+++ DataClasses/PointSubSet.py
@@ -20,7 +20,7 @@ class PointSubSet(PointSet):
             super(PointSubSet, self).__init__(points, path=path, intensity=intensity, range_accuracy=range_accuracy,
                                               angle_accuracy=angle_accuracy, measurement_accuracy=measurement_accuracy)
 
-        self.indices = indices
+        self.indices = indices.astype('int')
 
     @property
     def Size(self):
